move playersGrid to hook
